Move test clause to different file
